ventplotting/requirements.txt
scipy
numpy
pandas
h5py
matplotlib
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
coloredlogs
